remove useless option which epoch for test

+++ util/visualizer.py
@@ -27,13 +27,12 @@ class Visualizer():
             util.mkdirs([self.web_dir, self.img_dir])
         if opt.isTrain:
             self.log_name = os.path.join(opt.checkpoints_dir, opt.name, 'loss_log.txt')
-            with open(self.log_name, "w") as log_file:
-                now = time.strftime("%c")
-                log_file.write('================ Training Loss (%s) ================\n' % now)
+            if not os.path.isfile(self.log_name):
+                with open(self.log_name, "w") as log_file:
+                    now = time.strftime("%c")
+                    log_file.write('================ Training Loss (%s) ================\n' % now)
         else:
-            self.log_name = os.path.join(opt.results_dir, opt.name, '%s_%s.txt' % (opt.phase, opt.which_epoch))
-            with open(self.log_name, "w") as log_file:
-                log_file.close()
+            self.log_name = None # os.path.join(opt.results_dir, opt.name, '%s_%s.txt' % (opt.phase, opt.which_epoch))
 
     def reset(self):
         self.saved = False
